+++ app/build.gradle
@@ -144,7 +144,6 @@ ext {
             butterknife             : '8.7.0',
             firebase                : '9.4.0',
             stetho                  : '1.5.0',
-            chuck                   : '1.0.4',
             timber                  : '4.7.0',
             junit                   : '4.12',
             mockito                 : '2.18.3',
@@ -235,7 +234,6 @@ dependencies {
     // Instrumentation
     implementation "com.facebook.stetho:stetho:$dependencyVersions.stetho"
     implementation "com.facebook.stetho:stetho-okhttp3:$dependencyVersions.stetho"
-    //  debugImplementation "com.readystatesoftware.chuck:library:$dependencyVersions.chuck"
 
     // Firebase
     implementation "com.google.firebase:firebase-core:$dependencyVersions.firebase"
