+++ app/src/main/java/com/futurice/freesound/arch/viewmodel/SimpleDataBinder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Futurice GmbH
+ * Copyright 2016 Futurice GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,19 +14,24 @@
  * limitations under the License.
  */
 
-package com.futurice.freesound.mvi
+package com.futurice.freesound.arch.viewmodel;
 
-import android.arch.lifecycle.LiveData
+import android.support.annotation.NonNull;
 
-interface MviView<E : Event, in M : State> {
-    fun uiEvents(): LiveData<E> = EmptyLiveData()
-    fun render(model: M)
-    fun cancel() {}
+import io.reactivex.disposables.CompositeDisposable;
+
+/**
+ * Implements methods so that child classes don't need to declare empty implementations.
+ */
+public class SimpleDataBinder implements DataBinder {
+
+    @Override
+    public void bind(@NonNull final CompositeDisposable disposables) {
+        // Override this when required.
+    }
+
+    @Override
+    public void unbind() {
+        // Override this when required.
+    }
 }
-
-interface Renderer<in M : State> {
-    fun render(model: M)
-    fun cancel() {}
-}
-
-private class EmptyLiveData<T> : LiveData<T>()
