+++ app/src/main/java/com/futurice/freesound/arch/viewmodel/viewholder/BaseBindingViewHolder.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.futurice.freesound.viewmodel.viewholder;
+package com.futurice.freesound.arch.viewmodel.viewholder;
 
-import com.futurice.freesound.viewmodel.DataBinder;
-import com.futurice.freesound.viewmodel.ViewModel;
+import com.futurice.freesound.arch.viewmodel.DataBinder;
+import com.futurice.freesound.arch.viewmodel.ViewModel;
 
 import android.support.annotation.NonNull;
 import android.view.View;
