+++ app/src/main/java/com/futurice/freesound/feature/home/user/HomeUserUi.kt
@@ -18,7 +18,7 @@ package com.futurice.freesound.feature.home.user
 
 import com.futurice.freesound.feature.common.streams.Fetch
 import com.futurice.freesound.feature.common.streams.Operation
-import com.futurice.freesound.mvi.*
+import com.futurice.freesound.arch.mvi.*
 import com.futurice.freesound.network.api.model.User
 
 //typealias HomeUserFragmentViewModel = BaseViewModel<HomeUiEvent, HomeUiAction, HomeUiResult, HomeUiModel>
