+++ app/src/main/java/com/futurice/freesound/feature/search/SearchActivity.kt
@@ -22,7 +22,6 @@ import android.os.Bundle
 import android.support.v7.widget.SearchView
 import android.support.v7.widget.SearchView.OnQueryTextListener
 import android.view.View
-import android.widget.Button
 import com.futurice.freesound.R
 import com.futurice.freesound.app.FreesoundApplication
 import com.futurice.freesound.common.rx.plusAssign
@@ -31,9 +30,9 @@ import com.futurice.freesound.common.utils.ifNull
 import com.futurice.freesound.core.BindingBaseActivity
 import com.futurice.freesound.feature.common.scheduling.SchedulerProvider
 import com.futurice.freesound.inject.activity.BaseActivityModule
-import com.futurice.freesound.viewmodel.DataBinder
-import com.futurice.freesound.viewmodel.SimpleDataBinder
-import com.futurice.freesound.viewmodel.ViewModel
+import com.futurice.freesound.arch.viewmodel.DataBinder
+import com.futurice.freesound.arch.viewmodel.SimpleDataBinder
+import com.futurice.freesound.arch.viewmodel.ViewModel
 import io.reactivex.Observable
 import io.reactivex.ObservableEmitter
 import io.reactivex.Scheduler
