+++ app/src/main/java/com/futurice/freesound/network/api/FreeSoundApiService.java
@@ -37,4 +37,7 @@ public interface FreeSoundApiService {
 
     @NonNull
     Single<SoundSearchResult> search(@NonNull String query);
+
+    @NonNull
+    Single<SoundSearchResult> sounds(@NonNull String username);
 }
