--- app/src/main/java/com/futurice/freesound/viewmodel/BaseViewModel.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2016 Futurice GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.futurice.freesound.viewmodel;
-
-import android.support.annotation.NonNull;
-
-import io.reactivex.disposables.CompositeDisposable;
-
-public abstract class BaseViewModel implements ViewModel {
-
-    private final CompositeDisposable dataDisposable = new CompositeDisposable();
-
-    @Override
-    public final void bindToDataModel() {
-        bind(dataDisposable);
-    }
-
-    @Override
-    public final void unbindDataModel() {
-        dataDisposable.clear();
-        unbind();
-    }
-
-    protected abstract void bind(@NonNull final CompositeDisposable disposables);
-
-    protected abstract void unbind();
-
-    @Override
-    public final void dispose() {
-        dataDisposable.dispose();
-    }
-}
